kafka-producer/requirements.txt
kafka-python-ng
six